made common enum
